+++ test-parts/09-api_frontend/parts/03-get_device_list.js
@@ -57,7 +57,6 @@ function handleRetrieve()
 		it("Supported Models Retrieved", function(done)
 		{
 			modelReferenceArray = testCacheFile.getSupportedModels();
-			expect(modelReferenceArray).to.exist;
 			arrayFunctions.testPopulated(modelReferenceArray);
 			arrayFunctions.testAllType(modelReferenceArray, 'object');