+++ test-parts/09-api_frontend/parts/06-check_node_array.js
@@ -54,10 +54,7 @@ function handleManufacturerProperties()
 		it("Manufacturer Property List Retrieved From Cache", function(done)
 		{
 			manufacturersArray = objectFunctions.getProps(nodeCacheObject);
-			
-			expect(manufacturersArray).to.exist;
 			arrayFunctions.testPopulated(manufacturersArray);
-			
 			done();
 		});